requirements.txt
streamlit
ultralytics
Pillow
numpy
requests
torch

